+++ packages/adapter-interfaces/src/adapters/Users.ts
@@ -1,4 +1,4 @@
-import { User, Totals } from "../types";
+import { User, Totals, UserInsert } from "../types";
 import { ListParams } from "../types/ListParams";
 
 export interface ListUsersResponse extends Totals {
@@ -7,7 +7,7 @@ export interface ListUsersResponse extends Totals {
 
 export interface UserDataAdapter {
   get(tenant_id: string, id: string): Promise<User | null>;
-  create(tenantId: string, user: User): Promise<User>;
+  create(tenantId: string, user: UserInsert): Promise<User>;
   remove(tenantId: string, id: string): Promise<boolean>;
   list(tenantId: string, params: ListParams): Promise<ListUsersResponse>;
   update(tenantId: string, id: string, user: Partial<User>): Promise<boolean>;