+++ packages/adapter-interfaces/src/types/User.ts
@@ -28,7 +28,10 @@ export const userInsertSchema = baseUserSchema.extend({
   user_id: z.string().optional(),
   provider: z.string().default("email"),
   connection: z.string().default("email"),
+  is_social: z.boolean(),
 });
+
+export type UserInsert = z.infer<typeof userInsertSchema>;
 
 export const userSchema = z.object({
   ...userInsertSchema.shape,
@@ -36,7 +39,6 @@ export const userSchema = z.object({
   user_id: z.string(),
   // TODO: this not might be correct if you use the username
   email: z.string(),
-  is_social: z.boolean(),
   login_count: z.number(),
   identities: z.array(identitySchema).optional(),
 });